Documents/VM/ClientServiceDocs/client-worker/worker/requirements.txt
flask
flasgger
pyyaml
flask-swagger-ui
requests
kafka-python
